requirements.txt
numpy==1.26.4
pandas
scikit-learn
scikit-surprise
joblib
streamlit